backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pdfplumber
chromadb
openai
google-generativeai
sqlalchemy
psycopg2-binary
aiofiles
python-dotenv
requests
tqdm
pinecone
pydantic-settings
